Move init tests to a separate file
